refactor(extension): add swapRevamp flag, set up basic routes for the new version
