+++ apps/extension/src/app/features/feature-flags/index.tsx
@@ -30,6 +30,7 @@ interface FeatureFlags {
   releaseOnramperSell: boolean;
   assetsRevamp: boolean;
   activityRevamp: boolean;
+  swapRevamp: boolean;
 }
 
 export function useFlags() {
