+++ apps/extension/src/app/pages/swap-legacy/components/sbtc-ledger-recovery-warning.tsx
