+++ apps/extension/src/app/pages/swap-legacy/components/swap-asset-select/components/select-asset-trigger-button.tsx
