+++ apps/extension/src/app/pages/swap-legacy/components/swap-asset-select/components/swap-amount-field.tsx
