+++ apps/extension/src/app/pages/swap-legacy/components/swap-asset-select/components/swap-asset-selected.tsx
