+++ apps/extension/src/app/pages/swap-legacy/components/swap-assets-pair/swap-asset-item.layout.tsx
