+++ apps/extension/src/app/pages/swap-legacy/components/swap-assets-pair/swap-assets-pair.tsx
