+++ apps/extension/src/app/pages/swap-legacy/components/swap-details/swap-detail.layout.tsx
