+++ apps/extension/src/app/pages/swap-legacy/components/swap-details/swap-details.layout.tsx
