+++ apps/extension/src/app/pages/swap-legacy/components/swap-error.tsx
