+++ apps/extension/src/app/pages/swap-legacy/components/swap-review/bitcoin-swap-review.tsx
