+++ apps/extension/src/app/pages/swap-legacy/components/swap-review/stacks-swap-review.tsx
