+++ apps/extension/src/app/pages/swap-legacy/hooks/use-all-swappable-assets.tsx
