+++ apps/extension/src/app/pages/swap-legacy/hooks/use-bitcoin-bridge-assets.tsx
