+++ apps/extension/src/app/pages/swap-legacy/hooks/use-swap-navigate.ts
