+++ apps/extension/src/app/pages/swap-legacy/providers/stacks-swap-provider.tsx
