+++ apps/extension/src/app/pages/swap/swap-review.tsx
@@ -1,0 +1,5 @@
+import { Card } from '@app/components/layout';
+
+export function SwapReview() {
+  return <Card>{null}</Card>;
+}
