+++ apps/extension/src/app/pages/swap/swap.tsx
@@ -1,56 +1,19 @@
-import { Outlet } from 'react-router';
+import { Box } from 'leather-styles/jsx/index';
 
-import { SwapSelectors } from '@tests/selectors/swap.selectors';
-import { useFormikContext } from 'formik';
+import { RouteUrls } from '@shared/route-urls';
 
-import { Button } from '@leather.io/ui';
-import { isUndefined } from '@leather.io/utils';
+import { Card, Content, Page } from '@app/components/layout';
+import { PageHeader } from '@app/features/container/headers/page.header';
 
-import type { SwapFormValues } from '@shared/models/form.model';
-
-import { Card } from '@app/components/layout';
-import { LoadingSpinner } from '@app/components/loading-spinner';
-import {
-  SbtcLedgerRecoveryWarning,
-  isBtcToSbtcSwap,
-} from '@app/pages/swap/components/sbtc-ledger-recovery-warning';
-
-import { SwapAssetSelectBase } from './components/swap-asset-select/swap-asset-select-base';
-import { SwapAssetSelectQuote } from './components/swap-asset-select/swap-asset-select-quote';
-import { useSwapRouteParams } from './hooks/use-swap-route-params';
-import { type BaseSwapContext, useSwapContext } from './swap.context';
-
-export function Swap<T extends BaseSwapContext<T>>() {
-  const { isFetchingExchangeRate, isPreparingSwapReview } = useSwapContext<T>();
-  const { dirty, isValid, values } = useFormikContext<SwapFormValues>();
-
-  useSwapRouteParams();
-
-  if (isUndefined(values.swapAssetBase)) return <LoadingSpinner height="300px" />;
-
+export function Swap() {
   return (
-    <Card
-      dataTestId={SwapSelectors.SwapPageReady}
-      footer={
-        <Button
-          data-testid={SwapSelectors.SwapReviewBtn}
-          aria-busy={isPreparingSwapReview}
-          disabled={!(dirty && isValid) || isFetchingExchangeRate || isPreparingSwapReview}
-          type="submit"
-          fullWidth
-        >
-          Review and swap
-        </Button>
-      }
-    >
-      <SwapAssetSelectBase />
-      <SwapAssetSelectQuote />
-
-      {isBtcToSbtcSwap(values.swapAssetBase, values.swapAssetQuote ?? { tokenId: 'unknown' }) && (
-        <SbtcLedgerRecoveryWarning mt="space.05" />
-      )}
-
-      <Outlet />
-    </Card>
+    <Box width="100%">
+      <PageHeader title="Swap" onBackLocation={RouteUrls.Home} />
+      <Content>
+        <Page>
+          <Card>{null}</Card>
+        </Page>
+      </Content>
+    </Box>
   );
 }
